Fixed autocmd issue + :CrunchDev! replaces expr
Fixed the autocmd issues causing an error the first time CrunchDev is
called. This was caused by the autocmd code being in the autoload
directory, opposed to the plugin directory.

A bang can now be added to the end of the :CrunchDev command to replace
the expression with the result, opposed to the default behavior of
appending the result. Note: the above is reversed if the
g:crunch_result_type_append variable is set to 0. It is ! by default.

Also added some descriptions above a few functions.