fix e notation bug
Symptoms:
Variables using e notation that were in the operator's motion or visual
selection (captured' variables) to have their 'e' notation to be confused with a
variable name.

e.g.
When performing g= with the following 2 lines visually selected:

n = 2.304e7
n/10

This used to result in:

n = 2.304e7
n/10 = Crunch error: value for e7 not found

Problem:
The issue was that when the variable's value was incorporated back into the
expression before evaluation. An additional check was made for variables that
were not not in the visual selection or operator motion range, and were defined
somewhere above the lines being evaluated. This check was confusing the e
notation with a variable name.

Solution:
Account for the e notation like had been done before any variable replacing is
was done by replacing the e character in the new variable value with a #
character until all variable value replacement is done.

Fixes issue #52