fix issues with stale derived variable revaluation
Symptom:
When evaluating the entire buffer with the following:

Before:
---
x = 20
x = x = 10
x

After:
---
x = 20
x = x = 10
x = 10

Expected
---
x = 20
x = x = 20
x = 20

Problem:
This issues caused by The problem is that the value of 10 in the second
line was being captured as a variable and was overwriting the already captured
value of 20 in the first line.

Solution:
Only capture variables from lines that do not need revaluation when initially so
in the example above only the first line would have it's variable values
captured for later use.

fixes issues #42